C++ doesn't build in Ubuntu
git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1610798 13f79535-47bb-0310-9956-ffa450edef68

+++ lang/c++/impl/Compiler.cc
@@ -298,7 +298,7 @@ static GenericDatum makeGenericDatum(NodePtr n, const Entity& e,
             name = it->first;
             e2 = it->second;
         }
-        for (int i = 0; i < n->leaves(); ++i) {
+        for (size_t i = 0; i < n->leaves(); ++i) {
             const NodePtr& b = n->leafAt(i);
             if (nameof(b) == name) {
                 result.selectBranch(i);
